+++ NEWS.md
@@ -6,6 +6,7 @@
 - change smoothing in `scatterPlot` when `method = "level"` to use tensor interaction to allow for better smoothing when x and y are on different scales.
 - warn when >10% negative data detected in `polarPlot` and suggest setting `force.postive = FALSE`
 - make `maps` Suggests rather than Depends
+- refine check on whether bias correction is needed in `windRose`
 
 # openair 2.0-0